chore: merge upstream changes (#110)

+++ coregrind/m_aspacemgr/aspacemgr-common.c
@@ -177,7 +177,7 @@ SysRes VG_(am_do_mmap_NO_NOTIFY)( Addr start, SizeT length, UInt prot,
       fd = -1;
    res = VG_(do_syscall7)(__NR_mmap, (UWord)start, length,
 			  prot, flags, fd, offset, offset >> 32ul);
-#  elif defined(VGP_amd64_freebsd)
+#  elif defined(VGP_amd64_freebsd) || defined(VGP_arm64_freebsd)
    if ((flags & VKI_MAP_ANONYMOUS) && fd == 0)
       fd = -1;
    res = VG_(do_syscall6)(__NR_mmap, (UWord)start, length,
@@ -394,7 +394,11 @@ Bool ML_(am_get_fd_d_i_m)( Int fd,
    SysRes res = VG_(do_syscall2)(__NR_fstat, fd, (UWord)&buf);
 #endif
    if (!sr_isError(res)) {
-      *dev  = (ULong)buf.st_dev;
+      /*
+       * This gets compared to the value obtained by sysctl KERN_PROC_VMMAP.
+       * For some reson that only uses 32bits, so truncate this to match
+       */
+      *dev  = (UInt)buf.st_dev;
       *ino  = (ULong)buf.st_ino;
       *mode = (UInt) buf.st_mode;
       return True;
